Move CauseList containers
