+++ src/components/containers/CauseList/SliderSection/SliderSection.js
