+++ src/components/containers/CauseList/SliderSection/SliderSection.spec.js
